+++ src/main/java/com/ureca/snac/infra/config/TossPaymentProperties.java
@@ -1,0 +1,21 @@
+package com.ureca.snac.infra.config;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+
+/**
+ * application.yml에 정의된 토스페이먼츠 관련 설정 객체 바인딩
+ */
+@Getter
+@RequiredArgsConstructor
+@ConfigurationProperties(prefix = "payments.toss")
+public class TossPaymentProperties {
+
+    private final String url;
+
+    private final String successUrl;
+    private final String failUrl;
+
+    private final String secretKey;
+}
