+++ src/main/java/com/ureca/snac/infra/dto/request/TossConfirmRequest.java
@@ -1,0 +1,8 @@
+package com.ureca.snac.infra.dto.request;
+
+public record TossConfirmRequest(
+        String paymentKey,
+        String orderId,
+        Long amount
+) {
+}
